# backend/docs/swagger.yaml
basePath: /
definitions:
...
  httpserver.SwaggerMessage:
    properties:
      content: {}
      role:
        example: user
        type: string
    type: object
  httpserver.SwaggerMessagesRequest:
    properties:
      max_tokens:
        example: 1024
        type: integer
      messages:
        items:
          $ref: '#/definitions/httpserver.SwaggerMessage'
        type: array
      model:
        example: grok-chat-expert
        type: string
      stream:
        example: false
        type: boolean
    type: object
...
info:
  contact: {}
  description: Grok Build 与 Grok Web 多账号 API 网关。
  title: Grok2API
  version: "1.0"
paths:
...
  /Anthropic/messages:
    post:
      consumes:
      - application/json
      parameters:
      - default: "2023-06-01"
        description: Anthropic API version
        in: header
        name: anthropic-version
        required: true
        type: string
      - description: 请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpserver.SwaggerMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建 Anthropic Message
      tags:
      - Messages
...
securityDefinitions:
  BearerAuth:
    description: 使用 "Bearer g2a_xxx_xxx"。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"